# Conflicts:
#	app/src/main/java/com/android/go/sopt/winey/presentation/main/feed/detail/DetailActivity.kt
#	app/src/main/java/com/android/go/sopt/winey/presentation/main/feed/detail/DetailFeedAdapter.kt
#	app/src/main/java/com/android/go/sopt/winey/presentation/main/feed/detail/DetailViewModel.kt

+++ app/src/main/java/com/android/go/sopt/winey/data/source/FeedDataSource.kt
@@ -36,7 +36,7 @@ class FeedDataSource @Inject constructor(
         feedService.getFeedDetail(feedId)
 
     suspend fun postComment(
-        feedId: Long,
+        feedId: Int,
         requestPostCommentDto: RequestPostCommentDto
     ): BaseResponse<ResponsePostCommentDto> =
         feedService.postComment(feedId, requestPostCommentDto)
